refactor(llm): split OpenAiClient concrete agent to openai.rs
- Move OpenAiClient struct and impl to openai.rs
- Keep only LlmProvider trait in mod.rs
- Update mod.rs to pub mod openai; and re-export OpenAiClient
- No logic changes, only code organization
